Shrink hello-world bundle: split classic Renderer + lazy routing keywords
Cuts the hello-world smoke test from 243.30 KB / 77.32 KB gzip to
168.59 KB / 53.67 KB gzip — a 30.6% gzip reduction — while leaving the
classic v2-app-template essentially flat (+0.21 KB gzip from one extra
side-effect import).

Three changes, in order of impact:

1. **Lazy `-mount` and `-outlet` keyword registration.** Until now
   `resolver.ts` statically imported `mountHelper` and `outletHelper`,
   which transitively pulled `@ember/engine/instance`,
   `@ember/routing/-internals` (for `generateControllerFactory`), and the
   rest of the routing/engine graph into every bundle that uses
   `renderComponent`. Replace the static import with a
   `registerBuiltInKeywordHelper(name, helper)` registry on the resolver,
   and add a side-effect-only `syntax/register-routing-keywords.ts` that
   classic-app setup imports from `setup-registry.ts`. Bundles that don't
   pull in `setup-registry` (i.e. the hello-world that only uses
   `@ember/renderer`) drop ~138 KB of routing + ~7 KB of engine code.

2. **Split classic `Renderer` subclass into `classic-renderer.ts`.**
   Move `Renderer extends BaseRenderer`, `ClassicRootState`, the concrete
   `DynamicScope` class, and the `View` interface out of `renderer.ts`.
   Hoists the imports those carry — `OutletView`, `createRootOutlet`,
   `RootComponentDefinition`, `makeRouteTemplate`, `renderMain`,
   `guidFor`, `getViewElement`, `getViewId`, `dict`, `createCapturedArgs`,
   `EMPTY_POSITIONAL`, `curry` — out of the renderer-only bundle. Adds a
   `RootState` interface so `RendererState` can manage either kind
   without statically depending on classic code. `setup-registry.ts` now
   imports `Renderer` from `./classic-renderer`. The renderer entry
   re-exports the classic types so existing `from '.../renderer'`
   import sites keep working.

3. **Replace `RSVP.defer` in `renderSettled` with native Promise.**
   Standalone this didn't move the bundle (rsvp was reachable via other
   paths), but together with #1 it lets the hello-world bundle drop the
   62 KB rsvp shared chunk entirely — `@ember/engine`, `@ember/routing`,
   and `@ember/-internals/runtime/lib/ext/rsvp` were the remaining
   consumers, and #1 pulls those off the renderer-only path.

Verified: `lint:eslint`, `type-check:internals`, `type-check:types`,
`type-check:handlebars`, `test:node`, `test:blueprints`, classic
v2-app-template build, hello-world build, and a vite dev build of the
full test suite all pass. Browser tests will run in CI.
