+++ packages/@ember/-internals/glimmer/lib/syntax/register-routing-keywords.ts
@@ -1,0 +1,6 @@
+import { registerBuiltInKeywordHelper } from '../resolver';
+import { mountHelper } from './mount';
+import { outletHelper } from './outlet';
+
+registerBuiltInKeywordHelper('-mount', mountHelper);
+registerBuiltInKeywordHelper('-outlet', outletHelper);
